backend/requirements.txt
# FastAPI и основные зависимости
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# База данных
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Аутентификация и безопасность
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Валидация данных
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Очереди и кэширование
celery==5.3.4
redis==5.0.1

# HTTP клиент
httpx==0.25.2
aiohttp==3.9.1

# Работа с файлами
python-magic==0.4.27
Pillow==10.1.0

# Логирование
loguru==0.7.2

# Уведомления
aiosmtplib==3.0.1

# Утилиты
python-slugify==8.0.1
python-dateutil==2.8.2

# Разработка и тестирование
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Переменные окружения
python-dotenv==1.0.0 